Fix Solaris compilation
As detailed in Issue #57642, the `test-suite` doesn't compile on Solaris.

This patch fixes the compilation issues as explained there.

Tested on `amd64-pc-solaris2.11` (8 failures, all `fpcmp-target: Comparison
failed`), `sparcv9-sun-solaris2.11` (11 failures, same `fpcmp-target`
errors), and `x86_64-pc-linux-gnu` (no regressions).

Differential Revision: https://reviews.llvm.org/D133579

+++ MicroBenchmarks/ImageProcessing/Dither/main.cpp
@@ -69,7 +69,7 @@ void BENCHMARK_ORDERED_DITHER(benchmark::State &state) {
   int height = state.range(0);
   int width = state.range(0);
   int m = state.range(1);
-  int n = pow(m, 2);
+  int n = pow((double)m, 2);
   int *outputImage = (int *)malloc(sizeof(int) * height * width);
   int *temp = (int *)malloc(sizeof(int) * height * width);
